Vuex refactor for geoChart and related
